updatet to maven. unit test implemented

+++ src/main/java/pl/com/konrad/factorial/Main.java
@@ -7,12 +7,12 @@ public class Main {
     public static void main(String[] args) {
         for (int i = 1; i <= 8; i++) {
             System.out.print(i + " ");
-            System.out.println(Factorial.countRecursive(i));
+            System.out.println(Factorial.countRecursive(BigInteger.valueOf(i)));
         }
 
         for (int i = 1; i <= 8; i++) {
             System.out.print(i + " ");
-            System.out.println(Factorial.countIterative(i));
+            System.out.println(Factorial.countIterative(BigInteger.valueOf(i)));
         }
     }
 }
